Use master-alike create_transaction interface.
Make the interface to the slippage `create_transaction` and
`process_order` as it is on master, so that there is less of a chance
for breaking existing algorithms that override slippage.

Create a TradeBar instance in slippage simulate, which holds the `price`
and `volume` for use by both process_order and create_transaction.