Allow OHLCV in slippage models.
Fix bug where the SidView's asset was an integer.

Change BarData to take a simulation dt func instead of simulation
instance, since testing with new interface required creating a BarData
object to pass to slippage. (And it is easier to create a lambda that
provides a dt instead of instantiating or mocking out the
TradeSimulation object.)