ENH: Add writer for equity minute bcolz format.
Implement a writer for minute data into a format comprised of multiple
ctables, one for each individual equity, with a common 'index' shared by
all ctables where a given a dt maps to the same array index for all
equities and fields.

This format is pulled from the lazy-mainline/Q2.0 branch, with some
changes to the interface.

Add basic retrieval of values at a given dt to reader. Not yet used by
Zipline simulations, but added to support unit tests.