ENH: DataFrames and Panels in Q2.0 run method.
So that both examples 'stand-alone' and Quantopian research usage of
Zipline works with Q2.0

This patch favors making initial compatibility work with the existing
examples interface (and thus Quantopian research), over tuning the
structure and performance of the Panel etc.

While working on this, noticed that `history_window` and
`load_raw_arrays` overlap enough that they could be collapsed,
especially with the change to history to support arrays.