PERF: Prefetch daily history windows.
Use AdjustedArrayWindow to prefetch daily history windows, advancing the
window forward on each call for a new end date of the window.

Locally, on an algorithm with 6 daily windows this speeds up the daily
1d history from 300ms -> 200ms per day, saving about 33%.

The main bottlenecks this addresses is:
- Re-reading the data from the daily bar bcolz files every simulation
day.
- Remove need to do all adjustments and rounding every day.

Also, changed the `load_raw_arrays` in the Panel daily bar reader to
provide nans for missing days, to match the behavior of the bcolz
loader. (Discovered when running test_examples.)