PERF: If the window is one day, no adjustments.
Noticed in the case where the window function is used to fetch the
minutes for the current day to update the last slot in a '1d' history
window.

Further improvement can be made on this by prefetching the day's data
into a Float64Window instead and then iteratively getting the latest
value.

On local machine this cuts the time per day for an algo which does 3
bars of '1d' history for 6 algos by about 33%. (0.5 seconds -> 0.35 seconds)