ENH: Adds BenchmarkSource to make it easy to pass benchmark returns.
Makes TradingAlgorithm.run use the benchmark_return_source which which
was previously ignored.

Adds BenchmarkSource class which adapts a mapping of dt -> returns into
a source that emits BENCHMARK events.

This code is all going away in q2.0 but this small convenience should
help users with custom benchmark data until that happens.