BUG: Fix history crash at data start.
Fix a bug when history is called in a backtest with a start date near
the first trading day of the daily bars file, where load_raw_arrays
would hit a key error because data was requested which preceded dates
stored in the daily bar reader's calendar.

In the case where the first day is before the daily bar start, truncate
the query to days in the daily bar readers range, and then prepend nans.