BUG: Fix '1d' history for last day in slot.
Fix the following behavior for the last slot in history with a '1d'
freq.

- open
 - old
    Always used the open on the first minute, (9:31)
 - fix
    Uses first non-nan open.
- high
 - old
    The existence of any nan from market open to dt would propagate and
    make the result nan.
 - fix
    nans no longer affect high, unless there are only nans between
    market open and given dt
- low
 - old
    The existence of any nan from market open to dt would propagate and
    make the result nan.
 - fix
    nans no longer affect low, unless there are only nans between
    market open and given dt
- close
 - old
    The close at the given dt was always used. If the close was a nan at
    that dt, a nan would be returned.
 - fix
    The close at the given dt, unless that a value is a nan; in that
    case the most recent non-nan value is returned.

This patch's main intent is to fix the bug affecting results, however with
this fix there is a decent sized performance gain from cutting out the
calls to `all_minutes_for_day`, as well as cutting the retrieval of the
minute window from market_window -> current dt, which is order ~n^2 with
respect to how many minute values are accessed.

This implementation leaves some performance considerations on the table.
Particularly:
- The minute window access in each of the aggregation methods could be
fetched and stored in memory instead of going back to the bcolz tape on
every query.
- The minute windows could be structured so that they are accessed in
groups of assets. This implementation incurs the cost of getting the
minute position from the cache (and/or calculating) the number of bars
times the number of assets, when that retrieval could only be done once
per minute.

Also, adds test fixtures for bcolz minute data and the tempdir
dependency for the bcolz data, for use by the new test suite.