MAINT: Provide prefetch history loader for 1m.
Factor out USEquityHistoryLoader into a base class which can be
overridden to support daily or minute history.

In support of reading the minute bars for a range of market minute by
the loader from the reader, put the entire minute index for the minute
bar reader into memory so that the two indexes can be readily compared.
(Though we may want to consider a better way to store just the half day
 information, since the half day only accounts for ~1% of trading days.)

Change some test case sim params so that the tests are started far
enough from the dataset start so that history calls can be made.