MAINT: Provide prefetch history loader for 1m.
Factor out USEquityHistoryLoader into a base class which can be
overridden to support daily or minute history.

Use the minute version of the history loader in data portal minute
equity window history calls, in the same manner that the daily history
loader is used to pre-fetch history windows and apply adjustments as the
window is rolled.

Change some test case sim params so that the tests are started far
enough from the dataset start so that history calls can be made.