DEV: Fix some bugs around history windows and assets delisting
Rewrote the daily and minute tests for windows that straddle an asset
end date.

Added tests for getting daily history for multiple assets, some of
which have delisted.  (both for pure daliy and blended daily/minute)

Cleaned up some of the end dates for the mock data in history tests.