ENH: Filter undajusted window to market minutes.
Use market closes in minute reader to filter out non-market minutes, by
creating a 'blacklist' of minutes between the early close and would be
market close if the day were regular.

Added because the incoming minute history loader needs an interface to
query for the exact market minutes in a range from the reader.