BUG: Fix crash on repeated dt on daily aggregator.
The daily history aggregator should support calling the same dt in a
row, as long as the values are monotonically increasing.

e.g. calling the OHCLV methods with 9:31, 9:32, 9:32, 9:32 should be
supported. However, calling a repeated dt (9:32, 9:32) was causing
various crashes in each of the methods because the only condition being
checked before attempting to backfill with a window of data was 'was the
last visited dt equal to the current dt minus one minute'. On repeated
minutes this check is False, because the last_visited_dt is equal to the
current dt, which then resulted in a backfill where the next minute
after the current dt was being used as start of the window, which is
incorrect.

Fix by checking for the case where the last visited dt is equal to the
current dt, if it is, just use the value which was stored for that dt.