BUG: when limit price is crossed, fill at limit price (not market price)
For "non-marketable" limit orders (limit price has been crossed)
the final price must be the limit price while for "marketable" limit orders
the market price must be used (as before)
To disinguish between marketable and non-marketable limit order we can use
the following check:
if both open and close price are below/above (buy/sell) the limit price
the order is markettable
if open price is above (or below for sell) limit price and close price is
below (or above for sell) limit price, then the order is non-marketable

This patch doesn't cover the scenario where the crossing of the limit
price happens between the close price of the previous minute and the
open price of the next minute. But in my understanding the close price
of any minute must be almost identical to the open price of the following
minute, so I don't believe this is a real issue.