Modifies get_supplementary_field to only accept sids
To remove the additional work of having to call lookup_generic, which we
suggest not to use in internal code. This way we know we're always just
working with sids.

(There's also a case for manually handling Asset objects as well, but we
(can easily add that as a case within the method.)