resubmit: comment out one line in blotter.py
Initial submission modified an old copy of the file
The cancel_all_orders_for_asset call can induce an assertion failure in the case of a reverse split and delisting. See example below.
In fact, in running the test from January 2015 through November 26, 2015, the following scenario played out.

The stock was ordered on 8/25, 2 shares long
The stock experienced a reverse split on 8/25 with a ratio of 7, making the order amount become 2/7 = 0
The stock was halted on 11/25
The past_auto_close_date test for the asset evaluated to true (Quandl set its end date to 11/24)
The cleanup_expired_assets subroutine was invoked, calling cancel_all on the open order for Noranda
Since the order amount was 0, its status was filled, so it was not deleted
The “assert orders” line failed
If the process finds that the current market data set is missing information about a stock, the order for that stock should be invalidated, but this should be done in a fashion which does not break the execution of assert statements. In fact, though, the check on status of open in the open_orders list is redundant, and thus can be removed without loss of generality.
I did not rerun the test with the latest version of all files, but I see no changes to blotter or tradesimulation which would prevent this error from occurring.